refactor: 작품 분야 상수화 및 타입 따로 관리
