+++ src/constants/artworkFields.ts
@@ -1,0 +1,14 @@
+const ARTWORK_FIELDS = [
+  { name: '전체', value: 'ALL' },
+  { name: '회화', value: 'PAINTING' },
+  { name: '공예/조각', value: 'CRAFTSCULPTURE' },
+  { name: '드로잉', value: 'DRAWING' },
+  { name: '판화', value: 'PRINTMAKING' },
+  { name: '서예', value: 'CALLIGRAPHY' },
+  { name: '일러스트', value: 'ILLUSTRATION' },
+  { name: '디지털아트', value: 'DIGITALART' },
+  { name: '사진', value: 'PHOTOGRAPHY' },
+  { name: '기타', value: 'OTHERS' },
+];
+
+export default ARTWORK_FIELDS;
